chore: Move aws provider around for init plugin system
