+++ cmd/login.go
@@ -13,13 +13,13 @@ import (
 
 	lconfig "github.com/louislef299/knot/internal/config"
 	"github.com/louislef299/knot/internal/envs"
+	pecr "github.com/louislef299/knot/internal/providers/aws/ecr"
+	peks "github.com/louislef299/knot/internal/providers/aws/eks"
+	poidc "github.com/louislef299/knot/internal/providers/aws/oidc"
 	"github.com/louislef299/knot/internal/version"
 	laws "github.com/louislef299/knot/pkg/aws"
 	"github.com/louislef299/knot/pkg/dlogin"
 	los "github.com/louislef299/knot/pkg/os"
-	pecr "github.com/louislef299/knot/plugins/aws/ecr"
-	peks "github.com/louislef299/knot/plugins/aws/eks"
-	poidc "github.com/louislef299/knot/plugins/aws/oidc"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
