+++ internal/config/profile.go
@@ -6,8 +6,15 @@ import (
 	"io"
 
 	"github.com/louislef299/knot/internal/envs"
+	"github.com/louislef299/knot/pkg/provider"
 	"github.com/spf13/viper"
 )
+
+type Profile struct {
+	Name     string                `toml:"name"`
+	Provider string                `toml:"provider"`
+	Config   provider.ConfigSchema `toml:"config"`
+}
 
 var ErrNoProfilesFound = errors.New("there were no profiles found in config")
 
@@ -17,9 +24,8 @@ func ListProfiles(out io.Writer) error {
 		return ErrNoProfilesFound
 	}
 
-	for k, v := range v.AllSettings() {
-		fmt.Fprintf(out, "%s: %s\n", k, v)
+	for _, k := range v.AllKeys() {
+		fmt.Fprintf(out, "%s\n", k)
 	}
-	fmt.Fprint(out, "\n")
 	return nil
 }
