+++ internal/providers/plugins.go
@@ -1,0 +1,5 @@
+package providers
+
+func GetAvailableProviders() []string {
+	return []string{"aws"}
+}
